recommend/givutake/requirements.txt
fastapi
uvicorn
PyJWT
python-multipart
pydantic
python-dotenv